Restructure event routes, add login validation.
